refactor: refactor RelaySelector component and related files
