--- src/components/RelaySelector/index.new.tsx
@@ -1,9 +1,0 @@
-import { WsConnectStatus } from 'core/worker/type';
-
-export interface RelaySelectorProps {
-  wsStatusCallback?: (WsConnectStatus: WsConnectStatus) => any;
-  newConnCallback?: (conns: string[]) => any;
-  className?: string;
-}
-
-export function RelaySelector() {}
