+++ src/components/shared/Cascader/index.tsx
@@ -1,22 +1,23 @@
 import * as Popover from '@radix-ui/react-popover';
 import { groupBy } from 'lodash-es';
 import { PropsWithChildren, useMemo, useState } from 'react';
-import { Option } from './types';
+import { ICascaderOption } from './types';
 import { CascaderOption } from './option';
 
 export type CascaderProps = PropsWithChildren<{
   defaultValue?: string[];
-  options: Option[];
+  options: ICascaderOption[];
   displayRender?: (
     label: string[],
-    selectedOptions?: Option[],
+    selectedOptions?: ICascaderOption[],
   ) => React.ReactNode;
+  onChange?(value: string[]): void;
 }>;
 
 const defaultDisplayRender = (label: string[]) => label.join(' / ');
 
 export function Cascader(props: CascaderProps) {
-  const { defaultValue, options } = props;
+  const { defaultValue, options, onChange } = props;
   const [opened, setOpened] = useState(false);
   const [activeValue, setActiveValue] = useState<string[]>(defaultValue ?? []);
   const displayRender = useMemo(
@@ -25,7 +26,7 @@ export function Cascader(props: CascaderProps) {
   );
 
   const selectedOptions = useMemo(() => {
-    const selectedOptions: Option[] = [];
+    const selectedOptions: ICascaderOption[] = [];
     let cursor = options;
     for (const value of activeValue) {
       const option = cursor.find(option => option.value === value);
@@ -42,20 +43,21 @@ export function Cascader(props: CascaderProps) {
 
   const handleClick = (value: string[]) => {
     setActiveValue(value);
+    onChange?.(value);
     setOpened(false);
   };
 
   return (
     <Popover.Root open={opened} onOpenChange={setOpened}>
       <Popover.Trigger asChild>
-        <div className="h-10 w-full border border-gray-200 hover:border-brand rounded-lg cursor-pointer select-none transition-colors">
-          <div className="h-full w-full py-2 px-3 flex items-center">
+        <div className="h-10 w-full border border-gray-200 hover:border-brand rounded-lg overflow-hidden cursor-pointer select-none transition-colors">
+          <div className="h-full w-full py-2 px-3 flex items-center bg-surface-02 ">
             {displayRender(activeValue, selectedOptions)}
           </div>
         </div>
       </Popover.Trigger>
       <Popover.Content
-        className="w-[var(--radix-popover-trigger-width)] rounded-lg overflow-hidden bg-surface-02 outline-none shadow"
+        className="w-[var(--radix-popover-trigger-width)] rounded-lg overflow-hidden bg-surface-02 outline-none shadow z-50"
         sideOffset={5}
       >
         {Object.entries(optionsGroup).map(([group, options]) => (
